+++ src/lib/modlist/InstallProgressPopup.svelte
@@ -2,6 +2,7 @@
 	import Popup from '$lib/components/Popup.svelte';
 	import { invokeCommand } from '$lib/invoke';
 	import type { InstallProgress } from '$lib/models';
+	import { activeProfile, refreshProfiles } from '$lib/stores';
 	import { formatTime, shortenFileSize } from '$lib/util';
 
 	import { listen } from '@tauri-apps/api/event';
@@ -30,6 +31,7 @@
 				case 'done':
 					progress.totalProgress = 1;
 					progress.installedMods = progress.totalMods;
+					refreshProfiles();
 					setTimeout(() => {
 						open = false;
 					}, 250);