+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- More game icons
+
+### Fixed
+
+- Direct launch mode
+- Profile mod count not updating when a mod is installing
+-
+
 ## 0.8.10 (2024-10-11)
 
 ### Changed